pyproject.toml
[tool.poetry]
name = "image-colour-palette-generator"
version = "0.1.0"
description = ""
authors = ["MCDAngelo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-wtf = "^1.2.1"
wtforms = "^3.1.2"
numpy = "^2.0.1"
bootstrap-flask = "^2.4.0"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
scikit-learn = "^1.5.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.11.1"
pylint = "^3.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
